+++ alacritty/src/url.rs
@@ -1,10 +1,9 @@
 use std::cmp::min;
 use std::mem;
 
+use crossfont::Metrics;
 use glutin::event::{ElementState, ModifiersState};
 use urlocator::{UrlLocation, UrlLocator};
-
-use font::Metrics;
 
 use alacritty_terminal::index::{Column, Point};
 use alacritty_terminal::term::cell::Flags;