+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.5.0-rc2
+
+### Added
+
+- Separate search label for backward search
+
+### Fixed
+
+- Selection tracking in search without vi mode
+- Crash when resizing below search label length
+
 ## 0.5.0-rc1
 
 ### Packaging